--- packages/lms/src/lib/tts/index.ts
@@ -1,4 +1,0 @@
-import transform from './transform';
-import { generateAudio } from './generateAudio';
-
-export default { transform, generateAudio };
