+++ packages/tts/README.md
@@ -1,0 +1,13 @@
+# @mollify/tts
+
+This package contains the Elevenlabs text to speech (TTS) service. The @mollify/tts package provides a centralized location for defining and managing the TTS service.
+
+## Installation
+
+To install the @mollify/tts package with `npm`, simply run the following command:
+
+```bash
+npm install --save-dev @mollify/tts
+# or
+npm i -D @mollify/tts
+```
