+++ packages/tts/index.ts
@@ -1,0 +1,4 @@
+import transform from "./lib/transform";
+import { generateAudio } from "./lib/generateAudio";
+
+export { transform, generateAudio };
