+++ packages/tts/lib/index.ts
@@ -1,0 +1,4 @@
+import transform from "./transform";
+import { generateAudio } from "./generateAudio";
+
+export default { transform, generateAudio };
