+++ packages/tts/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@mollify/tts",
+  "version": "0.0.1",
+  "description": "Elevenlabs text to speech for @mollify",
+  "author": "Fermain <Fermain@users.noreply.github.com>",
+  "homepage": "https://github.com/Fermain/-Mollify#readme",
+  "license": "ISC",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "directories": {
+    "lib": "lib"
+  },
+  "files": [
+    "lib"
+  ],
+  "scripts": {
+    "build": "tsc"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Fermain/-Mollify.git"
+  },
+  "bugs": {
+    "url": "https://github.com/Fermain/-Mollify/issues"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.4"
+  }
+}
